split session
