+++ src/actions/session/setSession.ts
@@ -1,0 +1,6 @@
+import { Session } from "@wharfkit/session";
+import { Client } from "../../client";
+
+export const setSession = async (client: Client, session: Session | null) => {
+  client.state.setState({ session });
+};
